Move logic to subview
